# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/integrations/vietmap/reverse": {
      "get": {
        "operationId": "vietMapReverse",
        "parameters": [
          {
            "name": "latitude",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": -90,
              "maximum": 90,
              "example": 10.0452,
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": -180,
              "maximum": 180,
              "example": 105.7469,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Branch location resolved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Branch location resolved"
                    },
                    "data": {
                      "$ref": "#/components/schemas/VietMapLocationResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Yêu cầu không hợp lệ hoặc dữ liệu validation không hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Cookie phiên đăng nhập không hợp lệ, hết hạn hoặc thiếu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "User không có quyền hoặc không thuộc phạm vi chi nhánh hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "502": {
            "description": "VietMap provider unavailable"
          },
          "504": {
            "description": "VietMap provider timeout"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Reverse geocode a branch location with VietMap v4",
        "tags": [
          "vietmap"
        ],
        "description": "Reverse geocode a branch location with VietMap v4"
      }
    },
    "/integrations/vietmap/autocomplete": {
      "get": {
        "operationId": "vietMapAutocomplete",
        "parameters": [
          {
            "name": "text",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 2,
              "maxLength": 200,
              "example": "Đường 30 tháng 4",
              "type": "string"
            }
          },
          {
            "name": "focusLatitude",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": -90,
              "maximum": 90,
              "example": 10.0452,
              "type": "number"
            }
          },
          {
            "name": "focusLongitude",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": -180,
              "maximum": 180,
              "example": 105.7469,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Location suggestions retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Location suggestions retrieved"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/VietMapSuggestionResponseDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Yêu cầu không hợp lệ hoặc dữ liệu validation không hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Cookie phiên đăng nhập không hợp lệ, hết hạn hoặc thiếu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "User không có quyền hoặc không thuộc phạm vi chi nhánh hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "502": {
            "description": "VietMap provider unavailable"
          },
          "504": {
            "description": "VietMap provider timeout"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Search branch location suggestions with VietMap v4",
        "tags": [
          "vietmap"
        ],
        "description": "Search branch location suggestions with VietMap v4"
      }
    },
    "/integrations/vietmap/place": {
      "get": {
        "operationId": "vietMapPlace",
        "parameters": [
          {
            "name": "reference",
            "required": true,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 2048,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Place resolved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Place resolved"
                    },
                    "data": {
                      "$ref": "#/components/schemas/VietMapLocationResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Yêu cầu không hợp lệ hoặc dữ liệu validation không hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Cookie phiên đăng nhập không hợp lệ, hết hạn hoặc thiếu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "User không có quyền hoặc không thuộc phạm vi chi nhánh hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "502": {
            "description": "VietMap provider unavailable"
          },
          "504": {
            "description": "VietMap provider timeout"
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Resolve a VietMap v4 place reference",
        "tags": [
          "vietmap"
        ],
        "description": "Resolve a VietMap v4 place reference"
      }
    },
...
  },
  "info": {
    "title": "Bookora API",
    "description": "Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.",
    "version": "1.0",
    "license": {
      "name": "Proprietary"
    }
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "accessToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Cookie accessToken dùng để xác thực request cần user."
      },
      "refreshToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "Cookie refreshToken dùng cho refresh/logout phiên."
      },
      "csrfCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "csrfToken",
        "description": "Cookie csrfToken trong cơ chế CSRF double-submit. Phải đi cùng header X-CSRF-Token trên mutation cần CSRF."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Header CSRF double-submit. Giá trị phải trùng cookie csrfToken trên mutation cần CSRF."
      }
    },
    "schemas": {
...
      "VietMapLocationResponseDto": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "example": 10.0452
          },
          "longitude": {
            "type": "number",
            "example": 105.7469
          },
          "province": {
            "type": "string",
            "nullable": true,
            "example": "Thành phố Cần Thơ"
          },
          "ward": {
            "type": "string",
            "nullable": true,
            "example": "Phường Ninh Kiều"
          },
          "address": {
            "type": "string",
            "example": "12 Đường 30 tháng 4"
          },
          "displayAddress": {
            "type": "string",
            "example": "12 Đường 30 tháng 4, Phường Ninh Kiều, Thành phố Cần Thơ"
          }
        },
        "required": [
          "latitude",
          "longitude",
          "province",
          "ward",
          "address",
          "displayAddress"
        ]
      },
      "VietMapSuggestionResponseDto": {
        "type": "object",
        "properties": {
          "refId": {
            "type": "string"
          },
          "displayAddress": {
            "type": "string"
          }
        },
        "required": [
          "refId",
          "displayAddress"
        ]
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "required": [
          "statusCode",
          "message",
          "error",
          "path",
          "method",
          "timestamp"
        ],
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Dữ liệu không hợp lệ"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "code": {
            "type": "string",
            "example": "CSRF_INVALID"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "email": [
                "email must be an email"
              ]
            }
          },
          "path": {
            "type": "string",
            "example": "/api/v1/auth/login"
          },
          "method": {
            "type": "string",
            "example": "POST"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-28T12:00:00.000Z"
          }
        }
      }
    }
  }
}